# Fix am_delvariable() with dirty vars.