Fixes #15 by checking if any dirty variables exist in `am_delvariable()` and running the solver if so.

This introduces no overhead in the tight loop of `am_updatevars()`.