@szanni 2024-10-28
I am unsure how conclusive the `enaml_like_benchmark.cpp` benchmark is, the results seem to vary quite a bit on my system. Here are the results nonetheless:

a415445623ad66bd3fb737b161faf8702acfb6c4 (before any fixes):
```
|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|        1,522,523.08 |              656.80 |    0.1% |    3,341,504.64 |    3,206,382.67 |  1.042 |     818,019.40 |    7.1% |      0.18 | `building solver`
|            2,095.73 |          477,159.94 |    0.3% |       10,056.04 |        4,424.20 |  2.273 |       2,973.98 |    0.0% |      0.00 | `suggest value 400x600`
|            2,096.09 |          477,079.73 |    0.2% |       10,056.04 |        4,425.14 |  2.272 |       2,973.98 |    0.0% |      0.00 | `suggest value 600x400`
|            2,097.77 |          476,695.79 |    0.2% |       10,056.04 |        4,429.08 |  2.270 |       2,973.98 |    0.0% |      0.00 | `suggest value 800x1200`
|            2,073.28 |          482,326.50 |    0.7% |       10,056.04 |        4,377.39 |  2.297 |       2,973.98 |    0.0% |      0.00 | `suggest value 1200x800`
|            2,073.17 |          482,353.11 |    3.5% |       10,056.04 |        4,370.94 |  2.301 |       2,973.98 |    0.0% |      0.00 | `suggest value 400x800`
|            2,077.34 |          481,384.12 |    0.1% |       10,056.04 |        4,385.33 |  2.293 |       2,973.98 |    0.0% |      0.00 | `suggest value 800x400`

```

Current HEAD (9a164386cf66e3a85b0aad9c8b04ad7c6de2a329):
```
|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|        1,521,796.40 |              657.12 |    0.1% |    3,341,541.55 |    3,209,008.00 |  1.041 |     818,019.50 |    7.1% |      0.18 | `building solver`
|            2,069.71 |          483,160.44 |    3.4% |       10,130.03 |        4,369.80 |  2.318 |       2,998.98 |    0.0% |      0.00 | `suggest value 400x600`
|            2,005.14 |          498,718.29 |    0.7% |       10,130.03 |        4,235.44 |  2.392 |       2,998.98 |    0.0% |      0.00 | `suggest value 600x400`
|            2,093.57 |          477,653.00 |    0.2% |       10,130.03 |        4,417.09 |  2.293 |       2,998.98 |    0.0% |      0.00 | `suggest value 800x1200`
|            2,097.83 |          476,683.94 |    0.5% |       10,130.03 |        4,426.02 |  2.289 |       2,998.98 |    0.0% |      0.00 | `suggest value 1200x800`
|            2,090.27 |          478,407.20 |    0.8% |       10,130.03 |        4,413.01 |  2.295 |       2,998.98 |    0.0% |      0.00 | `suggest value 400x800`
|            2,104.63 |          475,142.54 |    2.2% |       10,130.03 |        4,438.45 |  2.282 |       2,998.98 |    0.0% |      0.00 | `suggest value 800x400`
```

This branch:
```
|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|        1,524,276.17 |              656.05 |    0.1% |    3,341,504.64 |    3,207,541.33 |  1.042 |     818,019.50 |    7.1% |      0.18 | `building solver`
|            2,055.77 |          486,436.60 |    0.1% |       10,056.04 |        4,339.34 |  2.317 |       2,973.98 |    0.0% |      0.00 | `suggest value 400x600`
|            2,055.83 |          486,420.75 |    0.2% |       10,056.04 |        4,338.49 |  2.318 |       2,973.98 |    0.0% |      0.00 | `suggest value 600x400`
|            2,073.09 |          482,372.26 |    0.1% |       10,056.04 |        4,378.39 |  2.297 |       2,973.98 |    0.0% |      0.00 | `suggest value 800x1200`
|            2,078.78 |          481,051.43 |    0.2% |       10,056.04 |        4,389.73 |  2.291 |       2,973.98 |    0.0% |      0.00 | `suggest value 1200x800`
|            2,073.70 |          482,230.84 |    0.2% |       10,056.04 |        4,376.08 |  2.298 |       2,973.98 |    0.0% |      0.00 | `suggest value 400x800`
|            2,075.50 |          481,812.74 |    0.2% |       10,056.04 |        4,382.91 |  2.294 |       2,973.98 |    0.0% |      0.00 | `suggest value 800x400`
```

General tendency being:

- before modification `> 480,000 op/s`
- current head `< 480,000 op/s`
- this patch `> 480,000 op/s`.