@starwing 2024-10-28
There is another way to reduce the performance gap, I'll make a branch tomorrow. Just link the dead variables in a list, and delete them after tight loop.