@starwing 2024-10-29
@szanni I have made a branch for dirty vars, could try it out to see the change of performance?

https://github.com/starwing/amoeba/tree/dirtyvar