@szanni 2024-10-29
> @szanni I have made a branch for dirty vars, could try it out to see the change of performance?
> 
> https://github.com/starwing/amoeba/tree/dirtyvar

Haven't studied the code in depth yet, but performance seems to be better (not sure how conclusive these numbers really are).

```
|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|        1,526,150.73 |              655.24 |    0.1% |    3,327,637.55 |    3,218,296.00 |  1.034 |     818,019.40 |    7.2% |      0.18 | `building solver`
|            2,035.89 |          491,186.71 |    0.5% |       10,044.03 |        4,291.50 |  2.340 |       2,960.98 |    0.0% |      0.00 | `suggest value 400x600`
|            2,026.16 |          493,545.29 |    0.7% |       10,044.03 |        4,279.10 |  2.347 |       2,960.98 |    0.0% |      0.00 | `suggest value 600x400`
|            2,046.07 |          488,741.94 |    0.5% |       10,044.03 |        4,319.65 |  2.325 |       2,960.98 |    0.0% |      0.00 | `suggest value 800x1200`
|            2,046.17 |          488,716.75 |    0.6% |       10,044.03 |        4,321.53 |  2.324 |       2,960.98 |    0.0% |      0.00 | `suggest value 1200x800`
|            2,046.96 |          488,530.16 |    0.1% |       10,044.03 |        4,323.96 |  2.323 |       2,960.98 |    0.0% |      0.00 | `suggest value 400x800`
|            2,047.62 |          488,372.66 |    3.6% |       10,044.03 |        4,324.11 |  2.323 |       2,960.98 |    0.0% |      0.00 | `suggest value 800x400`
```

If you're looking into getting even better performance: I would suggest looking into removing the bitfields as well (and making dirty a `_Bool`). On x86-64 I get consistently better results. Not too surprising I guess when considering it probably triggers a form of false sharing. It will cost an extra 8 bytes though for each `am_Var`.