@starwing 2024-10-30
Considering that the `am_Var` type only has a total of 48 bytes (on x86_64), the cost of increasing by 8 bytes is a bit too high. Could you provide the performance results after switching to _Byte, to see if this improvement is worth increasing the size of the `am_Var` type by 17%?