@starwing 2024-11-04
@szanni for the lack of follow-up I decide to merge branch `dirtyvar` into master, so if you have any good idea I will follow to change code :-)